requirements.txt
Django==5.2.5
python-dotenv
google-generativeai
